api/pyproject.toml
[tool.poetry]
name = "scorer-api"
version = "0.1.0"
description = "API for the Scorer application"
authors = ["Leandro Nicolás Acosta <leandroacosta97@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.26.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.1"
mangum = "^0.17.0"
pydantic = {extras = ["email"], version = "^2.6.0"}
loguru = "^0.7.2"
email-validator = "^2.1.0"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
passlib = "^1.7.4"
starlette = "^0.36.3"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 